segregate node mocks and browsersupport-node (RESEnvironment)
