+++ node/lib/browsersupport-node.js
@@ -1,22 +1,9 @@
-/* exported location, document */
+/* exported RESEnvironment */
 
-// Mocks
-
-var location = {"hash":"","search":"","pathname":"/","port":"","hostname":"www.reddit.com","host":"www.reddit.com","protocol":"https:","origin":"https://www.reddit.com","href":"https://www.reddit.com/","ancestorOrigins":{}};
-
-var document = {
-	location: location,
-	createElement: function(tagName) {
-		return {};
-	}
+var fs = require('fs');
+RESEnvironment.localStorageTest = function() { return true; };
+RESEnvironment.loadResourceAsText = function(filename, callback) {
+	callback(fs.readFileSync('lib/' + filename, 'utf8'));
 };
 
-var sessionStorage = {
-	getItem: function() {
-		return undefined;
-	}
-};
-
-exports.location = location;
-exports.document = document;
-exports.sessionStorage = sessionStorage;
+exports.RESEnvironment = RESEnvironment;
