+++ node/lib/mocks.js
@@ -1,0 +1,22 @@
+/* exported location, document, sessionStorage */
+// Mocks
+
+var location = {"hash":"","search":"","pathname":"/","port":"","hostname":"www.reddit.com","host":"www.reddit.com","protocol":"https:","origin":"https://www.reddit.com","href":"https://www.reddit.com/","ancestorOrigins":{}};
+
+var document = {
+	location: location,
+	createElement: function(tagName) {
+		return {};
+	}
+};
+
+var sessionStorage = {
+	getItem: function() {
+		return undefined;
+	}
+};
+
+exports.location = location;
+exports.document = document;
+exports.sessionStorage = sessionStorage;
+exports.DOMParser = require('xmldom').DOMParser;
